requirements/local.txt
-r base.txt
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-utils==0.3.0
async-factory-boy==1.0.1
pre-commit==2.20.0
psycopg2==2.9.9
ruff==0.1.2
SQLAlchemy-Utils==0.41.1
